+++ dev/test-harness/src/harness.rs
@@ -268,6 +268,7 @@ impl TestHarness {
             &self.notifier,
             cached_request,
             None,
+            false,
         )
         .await
         {